+++ CHANGELOG.md
@@ -13,7 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## 2019-10-05
 ### added
-- more tests for boot-gui-jpa
+- more tests for boot-gui-jpa, and spring-gui-jpa
 
 
 ## 2019-10-04